Fix sources: ["0"], output the correct source
Case 1, sourcemap missing
In [terser-js](https://github.com/terser-js/terser#source-map-options),
sources are always 0 if old sourcemaps are not provided.

The value passed for sourceMap.url is only used to set
//# sourceMappingURL=out.js.map in result.code.

The value of filename is only used to set file attribute
in source map file.

In broccoli-uglify-sourcemap we know in this case we are generating
sourcemap for the file we are processing, changing 0 to the actual
file gives us the correct source.

Case2, multiple sourceMap comments
source-map-url only matches the very first magic comment so we
mistakenly thinks the js file doesn't have a valid sourcemap.